Remove checks for zero variables.
The zero_* variables are generally in place for storing the data that is
of unknown origin and use (see http://stnsoft.com/DVD/index.html for
reading material). It's unused anywhere in libdvdread (and libdvdnav),
and so verifying the garbage data returns false positives.  It could be
a larger indicator that there are other issues with the IFOs, but not
enough to warrant any warnings.